+++ poem-openapi/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.2.57] 2022-2-10
+
+- Implement `From<T>`, `IntoIterator` for `MaybeUndefined<T>`.
+- Add `MaybeUndefined::from_opt_undefined`, `MaybeUndefined::from_opt_null`, `MaybeUndefined::as_ref`, `MaybeUndefined::as_deref` methods.
+
 # [1.2.56] 2022-2-10
 
 - Implement `ToJson` for `MaybeUndefined`.